backend/requirements.txt
# ===========================================
# SAURELLIUS CLOUD PAYROLL - Python Dependencies
# ===========================================

# Flask Framework
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.25

# Stripe Payments
stripe==7.10.0

# Email (Resend)
resend==2.0.0

# AI Integration
google-generativeai==0.8.0

# AWS (S3 for file storage)
boto3==1.34.25

# PDF Generation
playwright==1.41.0
pillow==10.2.0
qrcode==7.4.2

# Weather/Location APIs
requests==2.31.0
pytz==2024.1

# Utilities
python-dotenv==1.0.0
gunicorn==21.2.0
python-dateutil==2.8.2
Werkzeug==3.0.1

# Security
bcrypt==4.1.2

# Testing
pytest==7.4.4
pytest-flask==1.3.0

# Development
black==24.1.0
flake8==7.0.0
